requirements.txt
appdirs==1.4.4
asttokens==2.0.4
attrs==23.1.0
base58==2.1.1
bitarray==1.2.2
black==19.10b0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cytoolz==0.12.2
eth-abi==2.1.1
eth-account==0.5.2
eth-brownie==1.13.4
eth-event==1.2.3
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
execnet==2.0.2
hexbytes==0.2.1
hypothesis==5.41.3
idna==3.6
inflection==0.5.0
iniconfig==2.0.0
ipfshttpclient==0.7.0
jsonschema==3.2.0
lru-dict==1.3.0
more-itertools==10.1.0
multiaddr==0.0.9
mythx-models==1.9.1
netaddr==0.9.0
packaging==23.2
parsimonious==0.8.1
pathspec==0.11.2
pluggy==0.13.1
prompt-toolkit==3.0.8
protobuf==3.20.3
psutil==5.9.6
py==1.11.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pycryptodome==3.19.0
Pygments==2.6.1
pygments-lexer-solidity==0.5.1
PyJWT==1.7.1
pyrsistent==0.20.0
pytest==6.0.1
pytest-forked==1.6.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
pythx==1.6.1
PyYAML==5.3.1
regex==2023.10.3
requests==2.31.0
rlp==1.2.0
semantic-version==2.8.5
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
toolz==0.12.0
tqdm==4.53.0
typed-ast==1.5.5
urllib3==2.1.0
varint==1.0.2
vvm==0.1.0
vyper==0.2.16
wcwidth==0.2.12
web3==5.11.1
websockets==8.1
